Testing Bell's Inequality Locally
This file allows an individual to test Bell's inequality locally without using resources from a quantum computer. 

This is intended to let users get a more intuitive grasp of the simple circuit we are building (without the optimizations for the quantum computers), and to test if qiskit is properly installed on their device.

It utilizes information and code from IBM's CHSH Inequality Tutorial, without the code to optimize the circuit for quantum computations, aspects of the code changed to use StatevectorEstimator to simulate locally, and in-line annotations to give a clear understanding of what's going on.